requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.2.1
google-genai==1.63.0
Pillow==12.1.1
